pyproject.toml
[project]
name = "22-moana-alive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "jwt>=1.4.0",
    "ipykernel>=6.30.1",
    "fal-client>=0.7.0",
    "pydub>=0.25.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]
